Move tab and tabkey to domain
