+++ src/main/java/com/slayerassistant/rebuild/domain/TabKey.java
@@ -1,0 +1,20 @@
+package com.slayerassistant.rebuild.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum TabKey 
+{
+    LOCATIONS("Locations"),
+    ITEMS_REQUIRED("Items"),
+    COMBAT("Combat"),
+    MASTERS("Masters"),
+    WIKI("Wiki");
+    
+    private final String name;
+    TabKey(String name)
+    {
+        this.name = name;
+    }
+
+}
